PAYMENT SYSTEM - PYTHON
This program is a money payment system I made entirely in python. It works similar to the famous Mpesa system. When a user registers, their file is created and their info recorded in a general customer file.

Their file contains their name, a unique user id generated for them, the exact time they created their account and a list that is to be appended with their transactions both sent out and received money. All this info is stored in a json file.

The app also has a feature where the user requests their account activity to be sent to them via email which is pretty common and helpful. The program sends the details to them via mail!

Also, the json file that stores the user's account details has the ability to be queried for details, say, the user's transaction history.

Being a passionate python lover, I passionately built this program and it was a really good challenge to hone my skills further!